pursuit-docker/backend/pursuit_backend/requirements.txt
asgiref==3.5.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
certifi==2022.9.14
cffi==1.15.1
channels==3.0.5
charset-normalizer==2.1.1
constantly==15.1.0
cryptography==38.0.1
daphne==3.0.2
Django==4.1.1
django-cors-headers==3.13.0
django-environ==0.9.0
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
hyperlink==21.0.0
idna==3.4
incremental==21.3.0
mysqlclient==2.1.1
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.5.0
pyOpenSSL==22.1.0
pytz==2022.2.1
requests==2.28.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
Twisted==22.8.0
txaio==22.2.1
typing_extensions==4.3.0
urllib3==1.26.12
zope.interface==5.4.0
